Update conversion model to support big int field

+++ conversions/numbersapi/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 
 class Conversion(models.Model):
-    input_number = models.IntegerField()
+    input_number = models.BigIntegerField()
     output_words = models.CharField(max_length=200, blank=True)
     created_by = models.ForeignKey(User, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
